# [Performance]: Basic Benchmark & Multi-Put objects