--- a/pkg/dotc1z/resources.go
+++ b/pkg/dotc1z/resources.go
@@ -97,6 +97,39 @@ func (c *C1File) GetResource(ctx context.Context, request *reader_v2.ResourcesRe
 	}, nil
 }
 
+func (c *C1File) PutResources(ctx context.Context, resourceObjs ...*v2.Resource) error {
+	pairs := make([]*messageFieldsPair, 0, len(resourceObjs))
+	for _, resource := range resourceObjs {
+		updateRecord := goqu.Record{
+			"resource_type_id": resource.Id.ResourceType,
+			"external_id":      fmt.Sprintf("%s:%s", resource.Id.ResourceType, resource.Id.Resource),
+		}
+
+		if resource.ParentResourceId != nil {
+			updateRecord["parent_resource_type_id"] = resource.ParentResourceId.ResourceType
+			updateRecord["parent_resource_id"] = resource.ParentResourceId.Resource
+		}
+
+		pairs = append(pairs, &messageFieldsPair{
+			m:      resource,
+			fields: updateRecord,
+		})
+	}
+
+	err := c.db.WithTx(func(tx *goqu.TxDatabase) error {
+		err := c.bulkPutConnectorObjectQuery(ctx, tx, resources.Name(), pairs...)
+		if err != nil {
+			return err
+		}
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+	c.dbUpdated = true
+	return nil
+}
+
 func (c *C1File) PutResource(ctx context.Context, resource *v2.Resource) error {
 	updateRecord := goqu.Record{
 		"resource_type_id": resource.Id.ResourceType,