… go.sum
… vendor/github.com/doug-martin/goqu/v9/dialect/sqlite3/sqlite3.go
… vendor/modules.txt