- Adds a basic full Sync() benchmark.
- Based on early profiling of this, add a Multi-Write / Bulk `PutGrants()` function as an interface upgrade to Writer.  Just this change has a ~67.33% reduction in sync time for the basic benchmark.
- Also fix an issue with `goqu` Dialect;  We were trying to use `goqu.Dialect("sqlite3")`.  But if this Dialect isn't loaded, goqu makes a duplicate of the default... which is a huge memory allocation.  So... by using an underscore import of the dialect module, we cut memory allocations by ~16%.

```
goos: linux
goarch: amd64
pkg: github.com/conductorone/baton-sdk/pkg/sync
cpu: Intel(R) Xeon(R) E-2274G CPU @ 4.00GHz
               │   old.text   │              new.text               │
               │    sec/op    │   sec/op     vs base                │
ExpandCircle-8   1644.0m ± 1%   537.1m ± 2%  -67.33% (p=0.000 n=50)
```